backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
boto3>=1.34.0
requests>=2.31.0
Pillow>=10.0.0
pillow-heif>=0.16.0
apscheduler>=3.10.0
rapidfuzz
scikit-learn
numpy